requirements.txt
gpytorch==1.11
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.2
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.10.1
statsmodels==0.14.0
tensorflow==2.12.0
torch==2.0.1
botorch==0.11.0
